agent/requirements.txt
strands-agents>=1.23,<2.0
boto3>=1.35,<2.0
requests>=2.31,<3.0
